Move GlobStoreIdIterator implementation so we can have it non-public
