+++ libimagstore/src/store.rs
@@ -25,7 +25,7 @@ use walkdir::Iter as WalkDirIter;
 
 use error::{ParserErrorKind, ParserError};
 use error::{StoreError, StoreErrorKind};
-use storeid::{IntoStoreId, StoreId, GlobStoreIdIterator, StoreIdIterator};
+use storeid::{IntoStoreId, StoreId, StoreIdIterator};
 use lazyfile::LazyFile;
 
 use hook::aspect::Aspect;
@@ -34,6 +34,8 @@ use hook::accessor::{ MutableHookDataAccessor,
             StoreIdAccessor};
 use hook::position::HookPosition;
 use hook::Hook;
+
+use self::glob_store_iter::*;
 
 /// The Result Type returned by any interaction with the store that could fail
 pub type Result<T> = RResult<T, StoreError>;
@@ -1249,6 +1251,53 @@ impl Entry {
 
     pub fn verify(&self) -> Result<()> {
         self.header.verify()
+    }
+
+}
+
+mod glob_store_iter {
+    use std::fmt::{Debug, Formatter};
+    use std::fmt::Error as FmtError;
+    use glob::Paths;
+    use storeid::StoreId;
+
+    pub struct GlobStoreIdIterator {
+        paths: Paths,
+    }
+
+    impl Debug for GlobStoreIdIterator {
+
+        fn fmt(&self, fmt: &mut Formatter) -> Result<(), FmtError> {
+            write!(fmt, "GlobStoreIdIterator")
+        }
+
+    }
+
+    impl GlobStoreIdIterator {
+
+        pub fn new(paths: Paths) -> GlobStoreIdIterator {
+            GlobStoreIdIterator {
+                paths: paths,
+            }
+        }
+
+    }
+
+    impl Iterator for GlobStoreIdIterator {
+        type Item = StoreId;
+
+        fn next(&mut self) -> Option<StoreId> {
+            self.paths.next().and_then(|o| {
+                match o {
+                    Ok(o) => Some(o),
+                    Err(e) => {
+                        debug!("GlobStoreIdIterator error: {:?}", e);
+                        None
+                    },
+                }
+            }).map(|p| StoreId::from(p))
+        }
+
     }
 
 }
